Madisen

+++ src/src/DigitalIMU.cpp
@@ -77,9 +77,9 @@ void DigitalIMU::sample(IMUdata* data, State *ptr) {
     Serial.printf(" IMU_Quat_Y: %.5f\n",data->Quat[2]);
     Serial.printf(" IMU_Quat_Z: %.5f\n",data->Quat[3]);
 
-    ptr->quaternion(ptr->eulerAngle(data->GYRO[0], constants::gyro_sen, constants::dt), 
-                    ptr->eulerAngle(data->GYRO[1], constants::gyro_sen, constants::dt),
-                    ptr->eulerAngle(data->GYRO[2], constants::gyro_sen, constants::dt),
+    ptr->quaternion(ptr->eulerAngle(data->GYRO[0], constants::gyro_sen, (float)constants::dt), 
+                    ptr->eulerAngle(data->GYRO[1], constants::gyro_sen, (float)constants::dt),
+                    ptr->eulerAngle(data->GYRO[2], constants::gyro_sen, (float)constants::dt),
                     data->q_w, data->q_x, data->q_y, data->q_z);
 
     Serial.printf(" Quat_W: %.5f",data->q_w);
